Add /0 helper functions to point to a /1 function
I debated this a lot. In some ways, it seems redundant to have a
function with the same name just return another function. However, when
designing toward the final API, this makes some sense. For example

```elixir
Convention.new(:two_over_one, "Two Over One")
|> Bid.bid([Trait.opening_bid(), Level.one(), Strain.suit()], label: :opener)
|> Bid.pass()
|> Bid.bid([
  Level.two(),
  # Or Trait.non_jump_shift() + Strain.new() + Strain.suit()
  Strain.lt(Bid.previous_partner())
])
|> Convention.complete()
```

It would be weird for the second one to be `&Bid.pass/1` but the rest to
be ones that accept arguments. Additionally, for something like
`Strain.lt` it will need to accept an arg to work, but will still return
an anonymous function. Neverthless, this seems like something more
consistent where most "requirements/filters" take arguments that modify
and otherwise the user is ignorant of how they work.