Add pass as a possible generated bid
This adds pass as a possible generated bid. It is a somewhat hacky
solution in that a proper pass bid is {:pass, seat}, but the generator
could produce a bid that is {_, :pass, _} or {:pass, _, _}. However, as
long as those still fail it doesn't _really_ matter.

Additionally, since stream data is being used, we can basically
guarantee a {:pass, _} bid will be generated at some point, because the
generator will return that when the bid produced as {:pass, :pass, _}.