Add previous_partner function
This function will be passed to requirement that need access to a
specific bid. This and other functions like it (think of function where
we need a label for a bid) will examine the table and return a bid to
whichever requirement needs them.

The requirement functions that can accept this will know that they
receive a function or array of functions which will provide a specific
bid, or we can pattern match on the type.