Reverse order of strains
I plan to use the index of a strain in a strain list as the way of
determining higher/lower. As a result, the strains should go in order
from lowest to highest.

The levels already follow this pattern. While they are "numbers" they do
not actually have any traits of numbers (you can't do math on them) so
they will be used in a similar way.