Reorganize Structure, Add Models
Admittedly, being new to elixir this may not be a good idea.

However, it seemed necessary/useful to separate the model/structures of
levels and strains (maybe Bids?) as a separate module. Requirement
should always be functions that return functions that take a table. The
functions regarding structure and even helpers don't take a table and
thus really aren't a requirement to be used in a convention.

That distinction seems helpful, but for now, they are named the same,
that is the part that may be a bad decision. For now, we will see.