Add node object
A node represents a set of requirements (or a set of sets of
requirements) and a flag for how to process them (:bid or :one_of for
now).

The process method will figure out what kind of node it is, peel off a
bid from a table and apply all requirements. If they pass, it will
return {:ok, convention} or {:error, <something>}.