Add Support.Bids to enable better bid generation
Due to the way StreamData works, relying on it to generate a higher by
throwing away bids does not work. StreamData considers that to be a
failure. Instead of trying to generate a complicated building of higher
bids (which is a mix of any strain at a higher level or a higher strain
at the same level) is difficult in the stream data generation.

For the purposes of testing, we are just listing every possible contract
bid in order.