Convert bids to :ok, :next, :fail atoms
When processing conventions, we, in theory need to re-check every
convention after processing each bid. Something like a cue bid or
control bid may not happen right away, but it may happen later in the
hand.

This is fine, but we don't really want to check every convention if it
is no longer possible for the convention to be valid. Instead, we will
return :ok if a requirement is met, :next if the requirement was not
met, but continue checking the convention, and :fail if the requirement
was not met and therefore the convention cannot be met.

We can typically know this from the way in which the requirement fails.
I.e. if the requirement is that there is a 3 level bid, then any 1 or 2
level bid would be fine and we should keep lookign for that convention.
However if we have hit a 4 level bid and still never matched our
convention, we should stop checking.