Update processing to support :ok, :next, :fail
This is a large commit because it took me a long time to figure out.

All the initial attempts involved iterating through the requirements,
and if a requirement met a :next, then it would move onto the next bid
for the same requirement, but that logic is backwards.

What we need to do, is move through each bid in the table and check all
the requirements against a single bid. If even one requirement returns
:next, we move onto the next bid and try all of the requirements again.
I.e. all requirements must match for a single bid, not a single table.